requirements.txt
Django>=4.2
python-dotenv
requests
openai
gunicorn